Normalize order or type arguments for Query methods in Dispatcher and add XML documentation comments
