+++ Magneto.Version.props
@@ -2,6 +2,6 @@
 <Project>
 	<PropertyGroup>
 		<VersionPrefix>1.0.0</VersionPrefix>
-		<VersionSuffix>beta-6</VersionSuffix>
+		<VersionSuffix>rc-1</VersionSuffix>
 	</PropertyGroup>
 </Project>
